rfct: new file for sidebar components
